+++ src/App.js
@@ -6,6 +6,7 @@ import Selector, { filtersLoader } from "./components/Main/Selector/Selector.jsx
 import SelectorContextProvider from "./store/selector-context.jsx";
 import CompanyContextProvider from "./store/company-context.jsx";
 import User from "./components/Main/User/User.jsx";
+import RouteError from "./components/RouteError/RouteError.jsx";
 
 import "./App.css";
 
@@ -13,8 +14,9 @@ const router = createBrowserRouter([
   {
     path: "/",
     element: <Root />,
+    errorElement: <RouteError />,
     children: [
-      { index: true, element: <Metrics />, loader: thresholdsLoader },
+      { index: true, element: <Metrics />, loader: thresholdsLoader},
       { path: '/selector', element: <SelectorContextProvider><Selector /></SelectorContextProvider>, loader: filtersLoader},
       { path: '/company', element: <CompanyContextProvider><Company /></CompanyContextProvider>},
       { path: '/user', element: <User />},